Finishing podcast updater service

+++ Core/PodcastManager.Core.Domain/Models/PodcastStatus.cs
@@ -2,7 +2,7 @@ namespace PodcastManager.Domain.Models;
 
 public record PodcastStatus(
     DateTime NextUpdate,
-    int? TotalEpisodes = 0,
+    long? TotalEpisodes = 0,
     DateTime? LastTimeUpdated = null,
     int ErrorCount = 0,
     string[]? Errors = null
