+++ FeedUpdater/PodcastManager.FeedUpdater.Application.Tests/Doubles/EpisodeRepository/EpisodeRepositorySpy.cs
@@ -9,13 +9,13 @@ public class EpisodeRepositorySpy : EpisodeRepositoryStub
     public SpyHelper<(int, Item[])> SaveSpy { get; } = new();
     public SpyHelper<int> EpisodeCountSpy { get; } = new();
 
-    public override Task<(int, int)> Save(int code, Item[] feedItems)
+    public override Task<(long, long)> Save(int code, Item[] feedItems)
     {
         SaveSpy.Call((code, feedItems));
         return base.Save(code, feedItems);
     }
 
-    public override Task<int> EpisodeCount(int code)
+    public override Task<long> EpisodeCount(int code)
     {
         EpisodeCountSpy.Call(code);
         return base.EpisodeCount(code);
