+++ FeedUpdater/PodcastManager.FeedUpdater.Application.Tests/Doubles/EpisodeRepository/EpisodeRepositoryStub.cs
@@ -5,6 +5,6 @@ namespace PodcastManager.FeedUpdater.Application.Tests.Doubles.EpisodeRepository
 
 public class EpisodeRepositoryStub : EpisodeRepositoryDummy
 {
-    public override Task<(int, int)> Save(int code, Item[] feedItems) => Task.FromResult((1, 1));
-    public override Task<int> EpisodeCount(int code) => Task.FromResult(15);
+    public override Task<(long, long)> Save(int code, Item[] feedItems) => Task.FromResult((1L, 1L));
+    public override Task<long> EpisodeCount(int code) => Task.FromResult(15L);
 }
